chatbot-backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.78.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
